requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.1
marshmallow==3.14.1
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.1
marshmallow==3.14.1
Werkzeug==2.0.3
